[TASK] Update stack files to support PCIe based designs

+++ stack/src/kernel/dll/dllkcal.c
@@ -550,12 +550,12 @@ tOplkError dllkcal_asyncFrameReceived(tFrameInfo* pFrameInfo_p)
 #if CONFIG_DLL_DEFERRED_RXFRAME_RELEASE_ASYNC == FALSE
     // Copy the frame into event queue
     event.eventType = kEventTypeAsndRx;
-    event.pEventArg = pFrameInfo_p->pFrame;
+    event.pEventArg = (ULONGLONG)pFrameInfo_p->pFrame;
     event.eventArgSize = pFrameInfo_p->frameSize;
 #else
     // Only copy frame info into event queue
     event.eventType = kEventTypeAsndRxInfo;
-    event.pEventArg = pFrameInfo_p;
+    event.pEventArg = (ULONGLONG)pFrameInfo_p;
     event.eventArgSize = sizeof(tFrameInfo);
 #endif
 
@@ -619,7 +619,7 @@ tOplkError dllkcal_sendAsyncFrame(tFrameInfo* pFrameInfo_p,
     event.eventSink = kEventSinkDllk;
     event.eventType = kEventTypeDllkFillTx;
     OPLK_MEMSET(&event.netTime, 0x00, sizeof(event.netTime));
-    event.pEventArg = &priority_p;
+    event.pEventArg = (ULONGLONG)&priority_p;
     event.eventArgSize = sizeof(priority_p);
     ret = eventk_postEvent(&event);
